+++ CHANGES.md
@@ -1,7 +1,23 @@
 CHANGES
 ========
 
-## 3.1.7-RC
+## 3.1.8-RC
+
+* Feature: Login with Google Account
+* Feature: Login with GitHub Account
+* Feature: Attach files in Comment
+* Improvement: Write comment with CodeMirror Editor
+* Improvement: Post comment with Ctrl-Enter
+* Improvement: Place the commented page at the beginning of the list
+* Improvement: Resolve errors on IE11 (Experimental)
+* Support: Migrate to webpack 4 
+* Support: Upgrade libs
+    * eslint
+    * react-bootstrap-typeahead
+    * react-codemirror2
+    * webpack
+
+## 3.1.7
 
 * Fix: Update hidden input 'pageForm[grant]' when save with Ctrl-S
 * Fix: Show alert message when conflict